change from int to float to avoid mis-order frequencies

+++ src/emf_init_free.c
@@ -9,7 +9,7 @@
 #include "cstd.h"
 #include "emf.h"
 
-int cmpfunc(const void *a, const void *b) { return ( *(int*)a - *(int*)b ); }
+int cmpfunc(const void *a, const void *b) { return ( *(float*)a - *(float*)b ); }
 
 void emf_init(emf_t *emf)
 {
